# swagger.yaml
openapi: 3.0.0
info:
  title: Todo API
paths:
  /tasks:
    get:
      summary: Returns a list of all tasks in the database.
      responses:
        '200':
          description: A JSON array of tasks rows.
          content:
            application/json:
              schema: 
                type: array
                items: 
                  type: object
    post:
      summary: Create a new task and stores it in the database.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                title:
                  type: string
                description:
                  type: string
                completed:
                  type: boolean
      responses:
        '201':
          description: A message says Created.
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                type: string
  /tasks/{id}:
    put:
      summary: Updated a task with the given ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the task to update.
          schema:
            type: integer

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                completed:
                  type: boolean
      
      responses:
        '201':
          description: Task edited.
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Task not found.
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                type: string
    delete:
      summary: Deletes a task with the given ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the task to update.
          schema:
            type: integer
      
      responses:
        '201':
          description: Task deleted.
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Task not found.
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                type: string